+++ src/main/java/com/example/gastroValenciaApi/repositories/RestaurantRepository.java
@@ -3,7 +3,7 @@ package com.example.gastroValenciaApi.repositories;
 import com.example.gastroValenciaApi.models.RestaurantModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<RestaurantModel, Integer> {
+public interface RestaurantRepository extends JpaRepository<RestaurantModel, Long> {
 
 
 }
